ELEC-266: Moved scripts folder to x86
